Safari-611.1.10/LayoutTests/fast/dom/Window/open-window-min-size-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS 100 is >= 100
PASS 100 is >= 100
PASS 100 is >= 100
PASS 100 is >= 100
